Testing: add an additional informational message about QTextCodec selection
This will report whether a codec was actually generated (as `pCodec` will
not be a `nullptr` in that case - and then report the name it has. This
should clarify whether one was correctly generated for UTF-8 as opposed to
the previous one that indicated that a local8Bit one was being used in some
cases on Windows (as it was called "system") instead of the expected UTF-8.

Signed-off-by: Stephen Lyons <stephen@Hunt>