BugFix: prevent crashing from double-click on connect button
This is done by hiding the connect profile dialogue as quickly as possible
after the load or connect buttons have been clicked (and before the
possibly lengthy profile loading operations commenced). This should close
#5303.

Also, it was found that the:
`(void) dlgConnectionProfiles::slot_connectToServer()` slot method was
redundant and could be removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>