Enhance: get and set state of push-down-buttons (button love part1)  (#5227)
Provide a means to get and set the state of push down buttons (the latter
without executing the associate command or script). It requires adding
members to the `TAction` class to keep track of the `EAction` or
`TFlipButton` instance that is the implementation of the button on a
`TEasyButton` or `TToolBar` respectively. As it happens this showed that
the previous code to regenerate them was creating new buttons each time but
was not deleting the existing ones that were getting replaced.

Also:
* made some members of the `TAtion` class (the name, script and both
commands) private as otherwise there is no point in the setters and getters
also provided!
* simplified the existing `getButtonState()` {that only works within the
script for a button) to remove an unneeded intermediate local (`int`)
value.

Revised to combine ID and name getters and setters for button state

Note that for the getter this is an intermediate stage as is should be
possibble to combine it with the existing `getButtonState()` function that
takes NO arguments but only works within the Lua script for a push-down
button.

Revised to combine new getter functionality into existing getButtonState

This simplifies the UI so that `getButtonState()` works the same as
before but if an button's ID number or "name" is supplied as an argument
the state of the corresponding button (as a boolean though, NOT the
existing and, TBH, odd, integer value 1 or 2, that the original function
produces) is returned. In the event that the matching item is NOT a
push-down button, or that it does not match an existing button then a 'nil'
+ (run-time) error message is returned instead.

Revised to use ID consistently in two functions added/modified in this PR

Refactored to extract code to get TAction instance from {g|s}etButtonState(...)

This was strongly urged upon me by @keneanung & @vadi2 ...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>