BugFix: pass correct string length to zip_set_archive_comment(...)
The previous code was passing the length of a `QString` i.e. UTF-16BE
encoded data but the function needs to have the length of the string when
it is UTF-8 encoded (in bytes/chars) in a `QByteArray`. Whilst these values
will be the same if the string is purely an ASCII character one they will
be different as soon as the string contains characters outside of that
range of values.

It also repositions an adjacent comment that was in the wrong place because
of other additions to the same area of code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>