Enhance: improve zip comment generation for package exporter
Reproducing the exact content of the `config.lua` file is not exactly the
best it can be. This commit generates the comment text in a similar but not
identical manner to the existing config.lua file - but differs somewhat in
the textual presentation.

This might clash with #5356 as it happens to modify the same line as that
PR acts on to provide a bugfix - when resolving the conflict the line from
this commit MUST be used.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>