BugFix: make optional part of package exporter UI expand if there is space
This will close #5352.

The issue is that several `QLayouts` have been included on their own
instead of being assigned to a `QWidget` or other container widget. The fix
is to "morph" them into `QWidgets` which makes the later take on that
layout. In doing so I realised that where there were nested layouts (one
inside the other) there was some duplication of left/right margins so that
some of the displayed widgets were not as wide as some others; this was
improved by setting some of the left/right margins to zero.

It also corrects the spelling of a `QPushButton` pointer,
from: `pushBotton_packageLocation`
to: `pushButton_packageLocation`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>