Redo: get the numerus translations right for the package exporter & manager
Note that the "Remove package" text on the button in the package manager
has a similar issue in that it would update the text to say
"Remove package" or "Remove packages" in English (American) but if any
selection was cleared in use it would not update the text to say that NO
packages were currently selected for removal (though it would be disabled).

I have tweaked things so that the total number of packages to be removed is
now shown - even when it is zero. This will also work better for other
locales where zero might be handled differently to one (not though, in
France!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>