BugFix: another detail to try and fix odd NAWS messages
This seems to be successful as it eliminates NAWS messages when the screen
height is zero - which seems to be when the particular profile's main
`TConsole` is hidden during non multi-view operation when more than one
profile is loaded. The odd NAWS messages about windows of size 40 x 0 are
now explained as there is some code to maintain a minimum width of 40 for
the main `TConsole` - which is taking effect even in the hidden case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>