Update: add some more dictionaries to the list
Debian has added some which can show up in the lists for end-users.

For some reason it uses `gug` instead of `gn` for the Guarani language
which is used in Paraguay - so to cover it now and if they subsequently
correct it we need to code for both.

Also fix an error where an upper case country code was used for Turkish
when the comparison is made using entirely lower case strings.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>