Revise: following peer review of a duplicate PR
In #5415 which is a (presumably accidental) duplicate of this PR @vadi2
suggested the following as a replacement for the tool-tip for the option:
"Go-Ahead (GA) tells Mudlet when the game server is done sending text. On
games that don't support GA, this option controls how long Mudlet will wait
for more text to arrive. Larger values will help reduce the risk that a
large piece of text has unintended linebreaks in the middle of it, which
would break triggers. Lesser values will increase the risk of text getting
broken up, but will make the game feel more responsive."

in case it is not clear from the commit I have revised the text to partial
include some of that - and also taken on board Leris/Kebap who pointed out
on Discord that GA is not the only means of the Server signalling that it
has finished sending out packets. The text is now along the lines of
(within Markup limitations):
"*Go-Ahead* (`GA`) / *End-of-record* (`EOR`) signalling are Telnet enhancements
that tell Mudlet when the Game Server is done sending a piece of text. On
Game Servers which do not provide `GA` or `EOR` this option controls how
long Mudlet will wait for further network packets (i.e. text or other
'out-of-band' `OOB` data) to arrive. The default is 300 milli-seconds.
Larger values will help to prevent unintended linebreaks in the middle of
big pieces of text which may break triggers; smaller values may make Mudlet
seem more responsive to incoming text but runs a risk of breaking up text
(or more problematically `OOB` data).

*Adjustment of this persistent per-profile control is **NOT** recommended
unless you understand what it is doing and why it might be helpful.
Fine-tuning is likely to depend on particular network conditions and MUD
Game servers and a lower (shorter wait) setting that works for one user may
not be long enough for others.*"

He also suggested replacing the option as it is called
"Network packet timeout:"

on the basis that: "This text is great for an engineer who knows the
context, but not so much for a player who doesn't. How about:
'Wait up to [ ]ms for more text to arrive?'"

however I felt that was too wordy and could get broken in translation to
other locales. Also - given the potential for breakage,
particularly if it is set too small - we do not want the player who does
not get it, to play around with it.

Note that although I now mention something to the effect that the setting
is per profile and is saved between sessions that has not been coded yet.
It will follow in an additional commit that is pending.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>