Revise: finally nail the issue
It is because the problematic section of the form uses a `QFormLayout`
which is styled differently on MacOs (`QMacStyle`) compared to the styling
traditionally used for other platforms (`QCommonStyle`) - though modern
KDE desktops will use something a bit in between (`QPlastiqueStyle`) though
that still expands the widgets on the right side of a `QFormLayout`.

It should be possible to get the same behaviour on ALL platforms by
dropping back to a `QGridLayout`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>