BugFix: make optional part of package exporter UI expand if there is space (#5372)
* BugFix: make optional part of package exporter UI expand if there is space

This will close #5352.

The issue is that several `QLayouts` have been included on their own
instead of being assigned to a `QWidget` or other container widget. The fix
is to "morph" them into `QWidgets` which makes the later take on that
layout. In doing so I realised that where there were nested layouts (one
inside the other) there was some duplication of left/right margins so that
some of the displayed widgets were not as wide as some others; this was
improved by setting some of the left/right margins to zero.

It also corrects the spelling of a `QPushButton` pointer,
from: `pushBotton_packageLocation`
to: `pushButton_packageLocation`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: further tweaks to make some elements expand horizontally

Also fix another spelling mistake in `pushBotton_openInfos`
==> `pushButton_openInfos` .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: more vertical spacing tweaks in elements in packageExporter dialog

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Revise: finally nail the issue

It is because the problematic section of the form uses a `QFormLayout`
which is styled differently on MacOs (`QMacStyle`) compared to the styling
traditionally used for other platforms (`QCommonStyle`) - though modern
KDE desktops will use something a bit in between (`QPlastiqueStyle`) though
that still expands the widgets on the right side of a `QFormLayout`.

It should be possible to get the same behaviour on ALL platforms by
dropping back to a `QGridLayout`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>