BugFix: repair prevent all controls from being hidden feature (#5445)
It seems, *I think* that the removal of the "default host" profile (in
#2950) broke the logic that detected that any "real" profiles were loaded
and prevented both the main menu bar and the main toolbar from being hidden
when no such profiles were active. As it is now it is not possible to hide
both the main menu and the main toolbar AFTER a profile is loaded - even
though we provide the settings that are suppose to allow that.

The original `enum` used to encode the wanted settings for the visibility
of those two items used a cunning combination of bit patterns (and some
application of DeMorgan's laws) to work out what the visibility settings
are and what state the application is in (number of profiles active) so as
to show or hide the main menu and main toolbars. Unfortunately I've lost
track of how I put it together when the single "default host" profile
was included in the count so had to rewrite the logic tests without
changing the `enum` values.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>