Cleanup: remove some warnings in uses of QKeySequence
In code snippets like:
`inputLineKeySequence = QKeySequence(Qt::ALT | Qt::Key_L);`
this produces a warning:
"bitwise operation between different enumeration types ('Qt::Modifier' and
'Qt::Key1) [enum-enum-conversion]"

The cure is to use the correct constructor (having second to fourth
arguments defaulted to `(int) 0`) and not try to combine the two
different enumerations with a bitwise-OR:
`inputLineKeySequence = QKeySequence(Qt::Key_L, Qt::ALT);`

Also simplify a `QStringLiteral` used in the construction of a tooltip that
also carries a shortcut (i.e. a `QKeySequence`) indication.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>