Enhance: allow wait for more network packets to be adjusted (#5403)
This might be useful for tweaking for MUDs that do not use Go Aheads for
end-users with a good network connection.

In #5415 which was a (presumably accidental) duplicate of this PR @vadi2
suggested the following as a replacement for the tool-tip for the option:
"Go-Ahead (GA) tells Mudlet when the game server is done sending text. On
games that don't support GA, this option controls how long Mudlet will wait
for more text to arrive. Larger values will help reduce the risk that a
large piece of text has unintended linebreaks in the middle of it, which
would break triggers. Lesser values will increase the risk of text getting
broken up, but will make the game feel more responsive."

in case it is not clear from the commit I have revised the text to partial
include some of that - and also taken on board Leris/Kebap who pointed out
on Discord that GA is not the only means of the Server signalling that it
has finished sending out packets. The text is now along the lines of
(within Markup limitations):
"*Go-Ahead* (`GA`) / *End-of-record* (`EOR`) signalling are Telnet enhancements
that tell Mudlet when the Game Server is done sending a piece of text. On
Game Servers which do not provide `GA` or `EOR` this option controls how
long Mudlet will wait for further network packets (i.e. text or other
'out-of-band' `OOB` data) to arrive. The default is 300 milli-seconds.
Larger values will help to prevent unintended linebreaks in the middle of
big pieces of text which may break triggers; smaller values may make Mudlet
seem more responsive to incoming text but runs a risk of breaking up text
(or more problematically `OOB` data).

*Adjustment of this persistent per-profile control is **NOT** recommended
unless you understand what it is doing and why it might be helpful.
Fine-tuning is likely to depend on particular network conditions and MUD
Game servers and a lower (shorter wait) setting that works for one user may
not be long enough for others.*"

He also suggested replacing the option as it is called
"Network packet timeout:"

on the basis that: "This text is great for an engineer who knows the
context, but not so much for a player who doesn't. How about:
'Wait up to [ ]ms for more text to arrive?'"

however I felt that was too wordy and could get broken in translation to
other locales. Also - given the potential for breakage,
particularly if it is set too small - we do not want the player who does
not get it, to play around with it.

Note that although I now mention something to the effect that the setting
is per profile and is saved between sessions that has not been coded yet.
It will follow in an additional commit that is pending.

Revised the tool-tip text AGAIN and move the display of the unit (mSec)
to the label from the spin-box - to make translation a tiny bit simpler.

Also:
* To allow for the control to display the time in seconds it has to be
converted from a `QSpinBox` to a `QDoubleSpinBox` and some properties
tweaked to match - beyond the control the C++ code still uses an integer
representing the time out in milliseconds.
* Reorder the tab order for a couple of other controls that have been
added since it was last redone - otherwise tabbing across them will cause
odd jumps between different pages of the profile preferences dialogue.
* Remove a redundant default text for a label that "lights up" when
profile passwords are being migrated between secure and portable storage
modes - the text should never shown to the end-user - but just in case it
does a slightly more informative message than the default translatable
"TextLabel" will be used.

Revised the display of the unit from "mSec" to "seconds"in the spin-box

Signed-off by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>