Refactor: remove unneeded (and dangerously named) local variable
It turns out that the variable concerned, `__Pick` is a very badly named
(as symbols containing a double underscore or an underscore and followed by
an upper-case letter are explicitly reserved, e.g. see:
https://stackoverflow.com/a/1449301/4805858 )

Fortunately, it isn't needed and there is no need to pass the class member
it is a copy of via the `drawRoom(...)` method in the `T2DMap` class as
that can use the class value directly quite safely - as it never modifies
it.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>