do not trim leading whitespace for tab completion
Remove trimmed() from one spot in the tab completion code.  Look closer at it later and see if there is a reason it was there.  Trimmed does front and back, but back comes from the autocomplete dictionary thing so I -think- this is only consequence, unless the suggestion thing includes spaces at the end or something odd like that.  Issue 579.