Improve: simplify dlgRoomExits() invocation
This removes the need for the code that uses the `dlgRoomExits` class from
having to also call that class's `init()` method - the room ID for which
the dialogue is being used is now to be provided in the constructor rather
than in the now made private method.

As a side-effect it become obvious that there is no need to provide the
room ID as an argument to the `dlgRoomExit::initExit(...)` method as the
value is already available in the class's member `(int) mRoomID`!

This PR is a part that has been extracted from the draft (and to be
replaced) PR #5308 - that is to be broken up into smaller, easier to
digest, chunks! 8-)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>