Improve: update bundled NotoColorEmoji font
### This is for LINUX only!

This pulls in changes for Unicode 13.1. It also will remove older versions
of the font that prior versions of Mudlet will place in its own bit of the
file-system so that the FontManager class will not try to load more than
one instance of a font with an identical "family name" (which it will not
IIRC).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>