Improve: cleanup room_exits.ui file
This reorders the elements of two (nested as it happens) `QGridLayout`s so
that their elements are in ascending row then column order. It also
converts another one that only contains a single element to a `QVBoxLayout`.
This is so that a change to be made by another PR that will also be
extracted from the draft PR #5308 is easier to review.

This PR should not produce any functional changes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>