Improve: icons not coloured status text for normal exits in exits dialog
This PR puts an icon to the right of the text for each normal exit room ID
`QLineEdit` entry widget in the room exits form/dialogue. The icon is one
of four (blank, green tick, orange warning triangle, white-on-red cross)
to indicate whether the entry is:
* blank (empty)
* valid and to a room in the same area
* valid and to a room in a different area
* invalid
respectively. This replaces the previous use of coloured texts which did
not work well, particularly with dark styling.

This is a major functional part extracted from the previous draft mega-PR
#5303 and deals only with the normal exits. A separate PR is to follow to
deal with the special exits.

Also:
* removes predefined tooltips in the `room_exits.ui` for the normal exit
direction `QLineEdit` entry widgets - as the text is now incorrect and they
are not needed as the appropriate text is generated during the dialogue's
construction in the C++ code.
* refactors some very repeated code (one instance for each normal exit
direction) out of the:
  * `slot_????_textEdited(const QString&)`
  * `slot_stub_????_stateChanged(int)`
slot methods.
* removes a bogus tooltip that was assigned to the non-functional "key"
but which made mention of a southwest exit direction.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>