Improve: icons not coloured status text for special exits in exits dialog
This is a bit more involved than the normal exit ones as we have to provide
an extra class to be a "delegate" to handle the `QLineEdit` that is
embedded in the `QTreeWidget` derived special `ExitsTreeWidget` when a text
field (in this case the exit roomID) is edited. We use that and the code
from #5513 for normal exits to provide the same icons in that widget whilst
the exit roomID is being edited and we put the same icon in a new column to
the immediate right of that column when that field is not being edited.

This is a major functional part extracted from the previous draft mega-PR
#5303 and deals only with the special exit roomIDs. This follows on from a
separate PR #5513 already merged to handle normal exits.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>