Fix: enable restoration of editor toolbars even if both get hidden
This PR adds context menu actions to each of the `QTreeWidget` used for the
Mudlet items in the left side of the editor which will be shown if either
of the editor toolbars are hidden so that they can be restored even if both
get hidden.

This provides a similar UI experience to that in the main Mudlet
application window shown when both the main menu and the main toolbar get
hidden in that part.

This was prompted by a Discord user Kairuni (Keegan#5303) in the
#mudlet-development channel at 2021/10/20 19:14 UTC:
"Mm if I can find some time to add it, would we be open to a PR for adding
a 'reset editor bars' button, perhaps to the editor settings? Between being
able to drag them out of the window and turn them both off, it's amusingly
possible to lose both the items and actions bars and not be able to bring
them back without a restart"
"the workaround isn't too bad, just `lua saveProfile();` and restart
Mudlet"

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>