BugFix: undo a last minute change to "clean up" something that doesn't work
This is because there is no `QVection<QString>::append(QChar)` method, so
a `(enum) QChar::SpecialCharacter` has to be explicitly converted to a
`QString` *before* that can be appended to the container. I had forgotten
that and removed the intermediate `QString(...)` in what I thought was a
last minute improvement!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>