Fix: only provide NAWS information from main console (#5160)
I suspect that we might have been sending faulty information about
`TConsole`s other than the main console.

This might close #5154 by ensuring the values of `mScreenHeight` and
`mScreenWidth` from the upper pane of the `TMainConsole` are sent back to
the copies held in the `Host` class whenever they are modified as those are
the ones used in the NAWS messages.

This seems to be successful as it eliminates NAWS messages when the screen
height is zero - which seems to be when the particular profile's main
`TConsole` is hidden during non multi-view operation when more than one
profile is loaded. The odd NAWS messages about windows of size 40 x 0 are
now explained as there is some code to maintain a minimum width of 40 for
the main `TConsole` - which is taking effect even in the hidden case.

Revised to check and potentially send NAWS data on all main console
resizing: now stores the main console dimensions and only send NAWS
data if it changes. This is partially done by overloading the
resizeEvent() for that main console.

Also rename some methods to better reflect their functionality and remove
buffer `TConsole` instances from some of the
`TConsole::resizeEvent(QResizeEvent*)` code.

Revised to include check for Host pointer not being a nullptr!

This uses a slightly longer/more involved check than a prior effort
- hopefully it will be accepted when the last try just produced a more
lengthy warning message than before!

Revision by @atari2600tim - put update dimensions in
`activateProfile` (SlySven#41) not sure if this is ideal spot,
but this seems to do it!

Signed-off by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>
Co-authored-by: Tim Johnson <29287358+atari2600tim@users.noreply.github.com>