BugFix: update PCRE download source to one that still exists
The original "Official" site is no longer serving the old version 8.x
library up - instead they now point downloaders to an unoffical Source
Forge listing - it is not clear whether the particular URL I have put in
is a mirror that happened to be selected for me based on my geo-locaation,
this might not be optimum but it should at least work - to allow Windows
CI builds to resume.

As it happens this also brings in two minor version increments in the
library version from 8.43 to 8.45 though this library is now believed
to be frozen for any development.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>