Add: (Enhance) add map deletion functionality
This adds a `QPushButton` to allow the currently loaded/created map to be
completely removed. As a little protection against accident activation a
separate `QCheckBox` along side it must be checked first to enable it - a
bit like a guard over a physical toggle switch, a.k.a. a:
http://catb.org/~esr/jargon/html/M/molly-guard.html

It also adds a `deleteMap()` function (which takes no arguments) to the Lua
API - which returns `true` on success - or `nil` and an error message on
error.

This should close #5436 and #630, and thus:
https://bugs.launchpad.net/bugs/782843 from 2011.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>