fix buttontoolbar resizing on start (#5637)
#### Brief overview of PR changes/additions

When making syncing mpackage/xml modules asynchronous there was an issue with updating the button toolbar, as it can create issues (and is just wrong) to update gui elements from another thread.

I took a solution from https://stackoverflow.com/questions/21646467/how-to-execute-a-functor-or-a-lambda-in-a-given-thread-in-qt-gcd-style/21653558#21653558 to fix that problem by using a Timer set to 0 and executing in the main thread, but that seems to cause an issue when the toolbar has to be updated directly from the main thread.

So I decided to run the whole reloadModule function in the main thread to prevent this and other issues with calling gui elements from another thread.

This shouldn't make module syncing much slower (or hang mudlet) as the slowest functions (writing backups, zipping, writing module xmls) still run in another thread.

#### Motivation for adding to Mudlet
fix #5626 and a few other issues
#### Other info (issues closed, discussion etc)

#### Release post highlight
fix issue with button-toolbar on start

(cherry picked from commit bb4e1c83ebf9d63fa8fb5d180c303b53dd5d5c78)