Infrastructure (Public): replace QStringLiteral used for Tooltips
Specifically when it is used to put HTML like tags around the text so that it
triggers the "rich-text" detector in the Qt libraries so that the tips gets
better formatting than being a single (long) line of plain text.

Whilst working on this PR I also spotted and fixed:
* A couple of raw strings in the `VarUnit` class that should have been
translated (or replaced with a `QString()`) - this required adding the
`Q_DECLARE_TR_FUNCTIONS(`className`)` macro to that class as it is not
derived from the `QObject` class which we normally use to provide the
`QObject::tr(`...`)` method.
* The `XMLImport` class had an extrenuous ';' in the header file where
it had a `Q_DECLARE_TR_FUNCTIONS(XMLImport)` macro.
* The tooltip for the "Save item" button in the editor contained a
unmatched `<p>`...`</p>` tag pair as well as a useless (as it is ignored
there) `\n` {Line Feed}.
* The tooltip for the "Save profile" button in the editor contained a
unmatched `<p>`...`</p>` tag pair.
* I had previously defined the `QT_NO_CAST_FROM_ASCII` and
`QT_NO_CAST_TO_ASCII` macros in the `dlgRoomExits` class but since the
`mudlet.h` needs to be included for that class to give access to the
`mudlet::htmlWrapper(`...`)` that this PR now uses more widely that was
triggering warnings about raw string literals in *I think* third party
library code. So I had to remove those Qt macros.
* Generally, tool-tips are sentences so they should end with terminal
punctuation, usually a '.' and there were quite a few without those, they
have been fixed!
* The `mudlet::htmlWrapper(`...`)` has exactly the same code effect as
the `(const QString) singleParagraph` helper function in the `dlgRoomExits`
class - so it has been replaced by that.

This should close Issue #5691 but will cause #3952 to need tweaking to
at least remove a TODO that mentions that Issue!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>