Infrastructure (Public): replace QStringLiteral used for Tooltips (take 2)
This is a test version that implements #5694 a bit at a time so I can try
and deduce why the original is failing to CI using CMake - i.e. on pretty
much everything that is not the Window AppVeyor build.

This is intended to only be a DRAFT and not to be merged.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>