Improve variable icons by using a color which works in both light and dark modes (#5686)
#### Brief overview of PR changes/additions
The variables icons in dark mode are less than great. Ansi yellow, good ol' 128,128,0, looks pretty good against either white or black, so @ashira104 changed up the icons a bit to use that color, and this is what came out. I like it, but even if we decide to go with something different let's tackle this.