Fix: windowvista and Mactintosh styles not colouring PushButtons (again!)
This has come about because of #4984 (and then #5574) setting our own
style-factory before the orginal code (I put in as #4846) could determine
the original that Qt (or the user with a `--style xxxx` command line
argument) was going to use.

TODO: Add meaningful `objectName`s to the styles we impose that give an
indication of what has been selected. As it happens I got a big clue as to
why the `windowsvista` style was (again) not colouring the backgrounds of
`QPushButton`s was because the debug output that reported the "Style
factory" that was being used was consistently returning an empty string
(`""`) for that detail - and the #4846 code needed to see one of the
style-factories that were buggy in order to insert the workaround.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>