Improvement: expose cecho2ansi and have it work for non-ansi color names as well (#5672)
#### Brief overview of PR changes/additions
While perusing the ansi conversion functions I realized we were doing all the work of cecho2ansi but had not exposed it as separate functionality. Since we have done so for d and hecho, it made sense to me to perform this refactor.
In the process I also added support for colors which are not named ansi colors but exist in the color table. 
#### Motivation for adding to Mudlet
More complete API, consistency across the c/d/h variants.