Fix: get menu shortcuts and their key sequences working all the time again
This should close #3985, and *finally* #649! I was prompted to look into
this as a result of what @demonnic reported in:
https://github.com/Mudlet/Mudlet/pull/5715#discussion_r760282302

Previously the code was attempting to detect whether the main menu bar
itself was actually visible and using that to determine whether
EITHER: "key sequences" should be assigned to menu items and have the menu
items activate the desired slots (menu bar shown)
OR: "key sequences" should themselves be assigned to shortcuts that
directly call the desired slots (menu bar hidden)

This PR instead uses a state variable (boolean) and uses that to record
whether the menu bar is visible and only perform the changes required for
the above alternatives if the state is changed (or has not previously been
performed).

There have been several past attempts to get the short-cuts working:
* #2071 monitored the state of the main toolbar and NOT the menu bar
* #3994 changed to monitor the state of the menu bar but only enabled
(menu shown) or disabled (menu hidden) shortcuts that were assigned key
sequences and those shortcuts were connected directly to slots yet the key
sequences were assigned to menu items that themselves were ALSo connected
to the same slots.
* #5427 reverted #3994 so the code went back to that of #2071

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>