Infrastructure: refactor busted tests to help us determine code coverage (#5737)
#### Brief overview of PR changes/additions
Adjusts `describe` blocks in busted tests to fit a standard.
#### Motivation for adding to Mudlet
We don't have a great way to count code coverage given our API is written in two places. In our meeting today we discussed changing the way we categorize our tests in order to make it easier to run a grep against our test files to see what functions are represented by test cases or not.