BugFix: prevent problems with more than two profiles
There were two issues:
1. Opening a third profile when multi-view was activated did not continue
to show the second one. This was due to an unconditional hiding of the
profile that was active immediately before the new one was added. This
needed to be gated by the flag that multi-view was active - so that if that
was the case the previous profile was NOT then hidden.

2. When more than two profiles were present and multi-view was active
closing one of them was causing a segmentation fault. This was because a
`TMainConsole::resizeEvent(...) was being activated in the profile that was
being closed and at that point the base classes `TConsole::mpHost` was no
longer safe to use. The fix seemed to be to run the check for that pointer
being non-null BEFORE allowing the `QResizeEvent` to be passed from the
derived `TMainConsole` class's overriding event handler to the base classes
one - whereas it was being done AFTER in the existing code.

With the first of these two tweaks it was found that there is no need to
forcibly clear the "Multiview" setting when Mudlet reverts to only having
one profile active. Instead it can be left active - but ineffective - until
another profile is loaded, at which point the multi-view mode takes effect
again.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>