Fix: get replays working again
Add in some code I had to pre-parse ALL the replay data into an ordered
container (a `QMap<quint64, QByteArray>`) - the defect in the existing
replay code was that when I switched from use a `QTime` to a
`QElapsedTimer` in PR #4400 I neglected to account for the `elapsed()`
method's return type changing from an `int` (effectively a `int32_t`) to a
`qint64` (a.k.a. `int64_t`). This commit attempts to get around the issue
by trying to read the two integer values that preceed each "chunk" of
replay data in the recording first as two 4 byte `int32_t`s (actually
 `qint32`s) and if that does not work (because it produces obviously wrong
values - negative or overlarge) to try again with the first of the values
as an 8 byte `qint64`.

This commit also corrects the code that produces a recording file to use
a 4 byte `qint32` as the first value, restoring the format to that used
prior to the altered form in the original PR - which was introduced
***after*** Mudlet **4.10.1** was released.

TODO: Enhance the actual playback code so as to make use of the preparsed
recording data.

This should close #5570.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>